Adicionado rota para excluir um grupo e orgaizado funcoes de gerenciamento de grupos
