+++ urls.py
@@ -20,9 +20,10 @@ URLS = {
     URL("/api/v1/turmas/get"): view.get_turmas,
     URL("/api/v1/turmas/get_turma/:id"): view.obtem_turma_especifica,
     URL("/api/v1/turmas/:id/editar"): view.editar_turma,
-    URL("/api/v1/grupos/get"): view.get_grupos,
     URL("/api/v1/turmas/criar"): view.post_turma,
     URL("/api/v1/turmas/excluir/:id"): view.api_v1_turmas_excluir,
+    URL("/api/v1/grupos/get"): view.get_grupos,
+    URL("/api/v1/grupos/:id/excluir"): view.api_v1_excluir_grupos,
 }
